[ft #157963716]add the autoincrement of request id routes
